scmi: driver: modifies rx buffer size and timeout for rx response
Various platforms have different requirements for timings and message
size when processing the messages.

In case of Octeon TX, the time is for message processing is consiredably
higher than default set in the driver.

SCMI cpu-freq driver allows MAX_OPPS (16) OPPs. But when transferring
the 16 OPP data from SCP to ARM, because of max_msg_size limit,
it fails to transfer information about all 16 OPPs.
Increasing the max_msg_size prevents the problem to occur and ensures
correct data reception on Linux side.

Reviewed-on: https://sj1git1.cavium.com/19451 (timeout increase)
Reviewed-on: https://sj1git1.cavium.com/16842 (max_msg_size increase)
Change-Id: If2cb0ec7206bdbb9fb2695bab9c97e357a626d7f
Signed-off-by: Wojciech Bartczak <wbartczak@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/48454
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/115475
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>