octeontx2-af: suppress kpu profile loading warning
Currently there are three way in which a KPU profile can be
loaded (in high to low priority order):
1. profile image integrated in kernel image
2. firmware database method
3. default profile

In most of the cases profile is loaded using 2nd method, resulting
in a warning from linux firmware subsystem (method 1) due to absence
of firmware in kernel image.
Using an appropriate firmware_request API which can suppress such
warnings if no image is integrated to kernel image.

Change-Id: I86bf1bd1b4f389e15961d0a9ae22ce57f81097e0
Signed-off-by: Harman Kalra <hkalra@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/76236
Reviewed-by: Jerin Jacob Kollanukkaran <jerinj@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>