mm/vmstat: Do not queue vmstat_update if tick is stopped
From the vmstat shepherd, for CPUs that have the tick stopped, do not
queue local work to flush the per-CPU vmstats, since in that case the
flush is performed on return to userspace or when entering idle. Also
cancel any delayed work on the local CPU, when entering idle on nohz
full CPUs. Per-CPU pages can be freed remotely from housekeeping CPUs.

Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Bharat Bhushan <bbhushan2@marvell.com>
[Fix backporting issues]

Change-Id: Ic2755d527282e25e4001987a08a1aa8c44b8118c
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/87320
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/89283
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>