cn10k: Add workaround for MRVL TAD Errata-38891
An issue exists whereby if one core (core A) is accessing a cache
line and a certain set of race cases occur when a second core
(Core B) is executing a DC CVAC, DC CVAP or DC CVADP to the same
cache line, the second core B may respond to the Last Level Cache
(TAD's) request for data with an unexpected transaction type. This
will cause the TAD to ignore the response, response will never come
to Core A, so as soon as Core A reaches an instruction that requires
that response to complete it will hang. Core B and other cores will
remain responsive.

Workaround is to replace below cache maintenance instructions:
1. DC CVAC with DC CIVAC.
2. DC CVAP with DC CIVAC; DC CVAP.
3. DC CVADP with DC CIVAC; DC CVAP (DC CVAP is equivalent to
                                    CVADP for our system).

Any userspace execution of these instruction are configured to
trap to kernel and replace as above. Kernel usages of these
instruction are replaced using alternatives and compile time
kernel config CONFIG_ARM64_WORKAROUND_MARVELL_38891 option

Change-Id: I9230311bf9dc5a3ba114373b29c0835a9fae82ea
Signed-off-by: Bharat Bhushan <bbhushan2@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/93537
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/119887