arm_mpam: resctrl: Add support for mbm total
Currently, the driver enables "mbm_local" if mpam class that is picked
to be "L3" resource control, and supports the bandwidth counter enumerated
as part of "L3" resource control. But this will be a problem when the
bandwidth counter is hosted by DDR MSC, and is enumerated as part of
"MB" mpam class.

To support this configurtaion, lets enabled mbm_total if the mpam class
picked to be "MB" resource control. mbm_local is still used for the
bandwidth counter enumerated as part of "L3" resource control.

Signed-off-by: Amit Singh Tomar <amitsinght@marvell.com>
Change-Id: Ida7c0367465ba9f1343af3226f475b862d0fa214