arm_mpam: add support for MSMON_MBWU_L/CAPTURE
In MPAM v0.1 and from MPAM v1.1, There is an optional long monitor
register, MSMON_MBWU_L/CAPTURE, that contains a count of 44 bits
or 63 bits. Let's enable support for it.

Capture register MSMON_MBWU_CAPTURE that is loaded from the monitor
register each time the selected capture event occurs. When a capture
event occurs, the monitor register is copied to the capture register,
and this register read to get memory bandwidth utilization.

Signed-off-by: Amit Singh Tomar <amitsinght@marvell.com>
Change-Id: I0c7bb8d57a2f77e45b06df0529a5bf1624f66f77