octeontx2-pf: Fix check to identify if a ntuple rule has VF action
As of today, vf number returned from ethtool_get_flow_spec_ring_vf()
was used to identify if the ntuple rule has vf action. But that is not
correct since vf number is set to 255 for ntuple drop rule. This patch
fixes the same and uses a flag which is set only if the ntuple rule has
a valid vf action. Also, the patch added an elaborated comment on the
expected vf number based on different ethtool ntuple commands.

fixes: 486afa4357d6 ("octeontx2-pf: Allow ntuple rule to direct packet
to VF with higher Rx queue than its PF")
Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I4d1342f9d1744dfb442e91b13dceb216aa0f56e1