mm/vmstat: Use vmstat_dirty to track CPU-specific vmstat discrepancies
This patch will now use the previously introduced CPU-specific variable
namely vmstat_dirty to indicate if a vmstat differential/or imbalance is
present for a given CPU. So, at the appropriate time, vmstat processing can
be initiated. The hope is that this particular approach is "cheaper" when
compared to need_update(). The idea is based on Marcelo's patch [1].

[1]: https://lore.kernel.org/lkml/20220204173554.763888172@fedora.localdomain/

Signed-off-by: Aaron Tomlin <atomlin@redhat.com>
Signed-off-by: Bharat Bhushan <bbhushan2@marvell.com>
[Fixed some backporting issues]

Change-Id: Id46898559056f3f329f454f2ef4c704bf3260404
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/87319
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/89282