octeontx2-pf: Fix promisc mcam entry action
Current implementation is such that, promisc mcam entry action
is set as multicast even when there are no trusted VFs. multicast
action causes the hardware to copy packet data, which reduces
the performance.

This patch fixes this issue by setting the promisc mcam entry action to
unicast instead of multicast when there are no trusted VFs. The same
change is made for the 'allmulti' mcam entry action.

967db3529ec("octeontx2-af: add support for multicast/promisc packet
replication feature")
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Change-Id: I80f7eb687cacb98fe75aaba5e258ccacb3581ff8
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/115562
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>