octeontx2-af: fix CN10KB FIFO length for RPM2
CN10KB has RPM0 and RPM1 as 8 LMAC's and RPM2 with 4 LMAC's.
Similarly FIFO size of the RPM0 and RPM1 are 256KB each while
RPM2 is 128KB. With current code, while setting Tx credit
for MAX MTU, the FIFO length for RPM2 is coming to 16384
when it is divided by 8 and hence the tx_credit config
is becoming zero thereby breaking CN10KB's NIX Tx.

Change-Id: I7ec174fd80ee462ccdb83c00a475e9b6c888f72d
Signed-off-by: Nithin Dabilpuram <ndabilpuram@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/118280
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>