octeontx2-af: reveal only TIM params that are available
Due to the errata IPBUTIM-37450, TIM_AF_ADJUST_GTI register is
non-functional on T105B0 onward. Workaround is to not access the
register. So parameter to access the register is hidden on such HW.

If we don't conceal the devlink param, a read would return an error and
will cause "devlink dev param show" command to stop at that point.

Signed-off-by: Shijith Thotton <sthotton@marvell.com>
Change-Id: I955b48db5a431a4fc4a3638610fd58d8d09a0766
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/121736
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Pavan Nikhilesh Bhagavatula <pbhagavatula@marvell.com>
(cherry picked from commit 0943556194643327742d813c7d5039c199a697aa)
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/122084
Reviewed-by: Devapraba Muthumani <dmuthumani@marvell.com>