net: mvpp2: fix mvpp2 IRQ affinity issue
When CPU (cp1,cp2,cp3) is shutdown, mvpp2 driver IRQ
affinity is changed to available CPU's, which is not
restored when cpu is started again. This results into
getting prints "wrong cpu on the end of Tx processing",
while transmitting packets associated with that IRQ.
To resolve the issue, added a check to restore the
affinity of CPU by referring affinity_hint value before
transmitting any packets.

Signed-off-by: Anagha K J <aj@marvell.com>
Change-Id: Icbac0040c548eca4741557d1baf0bef8939a80c9
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/122461
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sanghoon Lee <salee@marvell.com>