arm64: dts: marvell: Fix pp2 uio nodes for cn9131/cn9132-db
Add/remove the definitions of CP11X_NUM before/after including
armada-cp115.dtsi. The missing definition leads the wrong uio
node name in device-tree for pp2 devices of the additional CPs,
such as:
  cp1/cp1:config-space@f4000000/f4000000.uio_pp_CP11X_NUM/uio/uio1
It causes that MUSDK failed to probe the Ethernet interface on CP1
and CP2 due to the mismatch of the node name.

Signed-off-by: Wilson Ding <dingwei@marvell.com>
Change-Id: Ie8642523f8f1ef214333d12470225dd91fc658f4
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/128049
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sanghoon Lee <salee@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/128361