octeontx2-bphy-netdev: handle cpri fifo overrun
netdev cannot distinguish between a full RX ring vs an empty
RX ring. This can cause that processing of CPRI UL ethernet
traffic to stop if the RX ring gets full.
This patch adds a timer based check for the fifo overrun
packets count and when it detects that the fifo overrun
counters is increasing and HW rd/wr pointers are the same
then it will schedule a napi poll to flush the RX ring.

Signed-off-by: Baha Mesleh <baha.mesleh@nokia.com>
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Change-Id: Ia21f4774b7c5297b18732b7928a00be823c22c4d
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/135018
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>