Octeontx2-pf: Perform NDC sync for NPA/NIX LFs
QOSS is designed to create a new send queue whenever  a class
is created, ensuringproper shaping and scheduling. However,
when multiple send queues are created and deleted in a loop,
SMMU errors are observed.

This patch addresses the issue by performing an NDC sync during
the teardown of QOS send queues.

Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Change-Id: I5bb298014a95b78ad3231338ce3a344dbd038e88
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/137147
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>