coresight: tmc-etr: Support larger trace buffers if reserved memory is insufficient
In the current reserved memory buffer allocation scheme, we cannot
allocate trace buffers larger than the reserved memory size. There are
use cases such as profiling via perf where we need to use large buffers.

So, add a check in tmc_etr_alloc_resrv_buf() for the requested buffer
size and fallback to other modes of allocation if the reserved memory
region is insufficient.

NOTE: When we fallback to other modes of allocation, preserved memory
tracing feature cannot be used.

Change-Id: I68410318f36e5ea97c1e41fe0ed68b9e024a23bf
Signed-off-by: Tanmay Jagdale <tanmay@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/137969
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Anil Kumar Reddy H <areddy3@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>