octeontx2-bphy-netdev: Fix race condition for timestamp packets
When multiple cores are enqueuing packets that require timestamping,
there can be a potential race condition where the netdev TX queue
will be stopped and there is no way to restart the TX queue to
resume the traffic from the TCP/IP stack. This patch fixes the race
condition by using synchronization mechanism when stopping and
restarting netdev TX queue.

Signed-off-by: Sai Krishna <saikrishnag@marvell.com>
Change-Id: I8661a0c83175450ccefcd2703cbf114d386878f2
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/138108
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>