arm64: Increase ARCH_FORCE_MAX_ORDER when Octeon RVU AF driver is enabled
The RVU mailbox HW implementation changed between CN10K and O20. On
O20, mailbox memory shared between RVU AF and the rest of the 95 RVU
PFs has to be a IOVA, which is translated by the SMMU to a PA. Due to
this fact, unlike CN10K where mailbox memory is reserved in the
firmware, the RVU AF driver needs to allocate and configure the
hardware.

Since there are 96 RVU PFs on O20, 96 * 64 kB = 6 MB (1536 pages with 4
kB page size) have to be allocated. By default, ARCH_FORCE_MAX_ORDER is
set to 11 which means that the maximum buffer size we can allocate is
2^10 pages = 1024, i.e. 4 MB with 4 kB page size. Since the RVU AF
driver is trying to allocate a 6 MB region, the allocation fails.

Change-Id: Ibe3ccdfdab3e7e00055cc9d55a8486acefb514a3
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/138144
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>